+++ apps/web/pages/auth/signin.tsx
@@ -1,17 +1,17 @@
-import type { GetServerSidePropsContext } from "next";
-import { getProviders, signIn, getCsrfToken } from "next-auth/react";
+import type { getProviders } from "next-auth/react";
+import { signIn } from "next-auth/react";
 
-import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
 import { Button } from "@calcom/ui";
 
 import PageWrapper from "@components/PageWrapper";
 
-type Provider = {
-  name: string;
-  id: string;
-};
+import { getServerSideProps } from "@server/lib/auth/signin/getServerSideProps";
 
-function signin({ providers }: { providers: Provider[] }) {
+function signin({ providers }: { providers: Awaited<ReturnType<typeof getProviders>> }) {
+  if (!providers) {
+    return null;
+  }
+
   return (
     <div className="center mt-10 justify-between space-y-5 text-center align-baseline">
       {Object.values(providers).map((provider) => {
@@ -29,21 +29,4 @@ signin.PageWrapper = PageWrapper;
 
 export default signin;
 
-export async function getServerSideProps(context: GetServerSidePropsContext) {
-  const { req } = context;
-
-  const session = await getServerSession({ req });
-  const csrfToken = await getCsrfToken(context);
-  const providers = await getProviders();
-  if (session) {
-    return {
-      redirect: { destination: "/" },
-    };
-  }
-  return {
-    props: {
-      csrfToken,
-      providers,
-    },
-  };
-}
+export { getServerSideProps };
